requirements.txt
beautifulsoup4==4.9.3
requests==2.26.0
urllib3==1.26.7
chardet==4.0.0
charset-normalizer==2.0.7
alembic==1.11.1
python-dotenv==1.0.0
psycopg2-binary==2.9.6
flask==2.3.2
pinecone-client==2.2.2
ndg-httpsclient==0.5.1
pyOpenSSL==23.2.0
pyasn1==0.5.0
flasgger==0.9.7b2
gunicorn==20.1.0
langchain==0.0.210
gevent==22.10.2
openai==0.27.8
faiss-cpu==1.7.4